third_party/blink/web_tests/animations/animation-css-rule-types.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<title>This test checks that the CSSRule RuleTypes for keyframe-related rules are what we expect</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<script>
var ruleType = window.CSSRule.KEYFRAMES_RULE;
assert_equals(ruleType, 7);
ruleType = window.CSSRule.KEYFRAME_RULE;
assert_equals(ruleType, 8);
done();
</script>
</body>
</html>
